Create a serai-primitives crate
Move types such as Amount/Coin out of validator-sets. Will be expanded in the
future.
