+++ substrate/serai/primitives/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "serai-primitives"
+version = "0.1.0"
+description = "Primitives for the Serai blockchain"
+license = "AGPL-3.0-only"
+repository = "https://github.com/serai-dex/serai/tree/develop/substrate/serai/primitives"
+authors = ["Luke Parker <lukeparker5132@gmail.com>"]
+edition = "2021"
+
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
+
+[dependencies]
+scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
+scale-info = { version = "2", default-features = false, features = ["derive"] }
+
+serde = { version = "1.0", features = ["derive"], optional = true }
+
+sp-core = { git = "https://github.com/serai-dex/substrate", default-features = false }
+sp-std = { git = "https://github.com/serai-dex/substrate", default-features = false }
+
+[features]
+std = ["scale/std", "scale-info/std", "serde", "sp-core/std", "sp-std/std"]
+default = ["std"]
