+++ substrate/validator-sets/pallet/Cargo.toml
@@ -18,6 +18,7 @@ scale-info = { version = "2", default-features = false, features = ["derive"] }
 frame-system = { git = "https://github.com/serai-dex/substrate", default-features = false }
 frame-support = { git = "https://github.com/serai-dex/substrate", default-features = false }
 
+serai-primitives = { path = "../../serai/primitives", default-features = false }
 validator-sets-primitives = { path = "../primitives", default-features = false }
 
 [features]
@@ -28,6 +29,7 @@ std = [
   "frame-system/std",
   "frame-support/std",
 
+  "serai-primitives/std",
   "validator-sets-primitives/std",
 ]
